+++ emscripten.py
@@ -754,6 +754,9 @@ class Memory():
     #  * then dynamic memory begins
     self.dynamic_base = align_memory(self.stack_high)
 
+    if self.dynamic_base >= shared.Settings.TOTAL_MEMORY:
+     exit_with_error('Memory is not large enough for static data (%d) plus the stack (%d), please increase TOTAL_MEMORY (%d) to at least %d' % (self.static_bump, shared.Settings.TOTAL_STACK, shared.Settings.TOTAL_MEMORY, self.dynamic_base))
+
 
 def apply_memory(js):
   # Apply the statically-at-compile-time computed memory locations.
@@ -2210,6 +2213,7 @@ def finalize_wasm(temp_files, infile, outfile, memfile, DEBUG):
     cmd.append('--separate-data-segments=' + memfile)
   if not shared.Settings.SIDE_MODULE:
     cmd.append('--initial-stack-pointer=%d' % Memory().stack_base)
+  shared.print_compiler_stage(cmd)
   stdout = shared.check_call(cmd, stdout=subprocess.PIPE).stdout
   if write_source_map:
     debug_copy(wasm + '.map', 'post_finalize.map')