+++ ChangeLog.md
@@ -15,6 +15,7 @@ full changeset diff at the end of each section.
 
 Current Trunk
 -------------
+- Add port for bzip2 library (`libbz2.a`). (#8349)
 
 v1.38.29: 03/11/2019
 --------------------